refactor: Move encode_next_offset into db module
